requirements.txt
feedparser>=6.0
edge-tts>=6.1
google-generativeai>=0.5
google-genai>=1.0
requests>=2.31
pyyaml>=6.0
notion-client>=2.2
